fix: cleanup post merge
