+++ .github/workflows/ci.yml
@@ -5,12 +5,6 @@ on: push
 
 jobs:
   cargo-lint:
-<<<<<<< HEAD
     uses: Cosmian/reusable_workflows/.github/workflows/cargo-nursery-base.yml@develop
     with:
       toolchain: nightly-2022-10-28
-=======
-    uses: Cosmian/reusable_workflows/.github/workflows/cargo-nursery.yml@main
-    with:
-      toolchain: stable
->>>>>>> c8dba97 (feat: get callback errors from Findex)
