+++ crates/cover_crypt/src/wasm_bindgen/mod.rs
@@ -4,18 +4,9 @@ macro_rules! wasm_unwrap {
     };
 }
 
-use wasm_bindgen::prelude::*;
-
 pub mod abe_policy;
 pub mod generate_cc_keys;
 pub mod hybrid_cc_aes;
-
-#[wasm_bindgen]
-extern "C" {
-    #[wasm_bindgen(js_namespace = console)]
-    fn log(s: &str);
-    fn alert(s: &str);
-}
 
 #[cfg(test)]
 mod tests;
