--- tests/findex/datasets/expected_db_uids.json
@@ -1,32 +1,0 @@
-[
-  4,
-  5,
-  7,
-  8,
-  14,
-  17,
-  19,
-  20,
-  23,
-  34,
-  37,
-  43,
-  46,
-  48,
-  55,
-  56,
-  60,
-  61,
-  63,
-  65,
-  68,
-  70,
-  71,
-  77,
-  80,
-  82,
-  83,
-  85,
-  86,
-  96
-]
