--- tests/findex/datasets/key.json
@@ -1,1 +1,0 @@
-6hb1TznoNQFvCWisGWajkA==
