--- tests/findex/datasets/label
@@ -1,1 +1,0 @@
-Some Label
